+++ front/pages/index.vue
@@ -6,7 +6,7 @@
 
 <script lang="ts">
 import { Component, Vue } from 'nuxt-property-decorator'
-import LoginForm from '~/components/LoginForm.vue'
+import LoginForm from '~/components/LoginForm'
 
 @Component({
   components: {
